@@ -1,3 +1,3 @@
 ﻿@using WebApp
-@using WebApp.Models
+@namespace WebApp.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
