@@ -1,3 +1,3 @@
-﻿@{
+@{
     Layout = "_Layout";
 }
